+++ draw.h
@@ -2,5 +2,5 @@
 #define _DRAW_H
 void drawSquare(GLfloat, GLfloat, GLfloat, GLfloat, GLfloat, GLfloat);
 void drawBlock1x1(GLfloat, GLfloat, GLfloat);
-void drawMap(int size, int map[size][size][size], GLfloat*, GLfloat*);
+void drawMap(int size, int map[size][size][size], GLfloat*, int*);
 #endif
